Implements Event creation

+++ src/main/java/com/vcp/hessen/kurhessen/data/event/EventRepository.java
@@ -2,18 +2,20 @@ package com.vcp.hessen.kurhessen.data.event;
 
 
 import com.vcp.hessen.kurhessen.data.User;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Repository
 public interface EventRepository
         extends
-            JpaRepository<Event, Long>,
+            JpaRepository<Event, Integer>,
             JpaSpecificationExecutor<Event> {
 
 
@@ -25,4 +27,6 @@ public interface EventRepository
             "   where p2.user.id = ?1"+
             "  )")
     List<Event> findEventsByParticipantsContainingUserId(Long userId);
+
+
 }
